Move IsInsideStructureTracker.java to CorgiLib
